Add styles and complete logic
