+++ src/modules/forms/plugins/Cors/CorsPlugin.js
@@ -14,7 +14,7 @@ import Hint from '../../../../components/Hint/Hint';
 import ControlBar from '../ControlBar/ControlBar';
 import MultiSelect from '../../../selects/MultiSelect/MultiSelect';
 import TagSelect from '../../../selects/TagSelect/TagSelect';
-import Control from '../../../../components/Control/Control';
+import CopyToClipboard from '../../../../components/CopyToClipboard/CopyToClipboard';
 
 const propTypes = {
     apiSchema: PropTypes.object.isRequired,
@@ -32,19 +32,8 @@ class CorsPlugin extends PureComponent {
         this.setState({});
     }
 
-    copyToClipboard = () => {
-        const fakeInput = document.createElement('textarea');
-
-        fakeInput.setAttribute('id', 'fakeInput');
-        fakeInput.value = getValues(['config', 'request_headers'])(this.props.pluginFromValues);
-        document.body.appendChild(fakeInput);
-        fakeInput.select();
-        document.execCommand('copy');
-        document.body.removeChild(fakeInput);
-    }
-
     render() {
-        const { apiSchema, className, edit, name, handlePluginExclude, plugin, pluginName } = this.props;
+        const { apiSchema, className, edit, name, handlePluginExclude, plugin, pluginFromValues, pluginName } = this.props;
         const b = block(className);
         const allValues = key => apiSchema.plugins[0].config[key];
 
@@ -88,19 +77,19 @@ class CorsPlugin extends PureComponent {
                 <Row className={b('row')()} fullwidth>
                     <Row col>
                         <Label>Request Headers</Label>
-                        <Control
-                            icon="copy"
-                            onClick={this.copyToClipboard}
-                        />
-                        <Field
-                            name={`${name}.config.request_headers`}
-                            type="text"
-                            edit={edit}
-                            value={() => getValues(['config', 'request_headers'])(plugin)}
-                            placeholder={SETUP.placeholders.cors.request_headers}
-                            options={optionsTransformer(allValues('request_headers'))}
-                            component={TagSelect}
-                        />
+                        <CopyToClipboard
+                            value={getValues(['config', 'request_headers'])(pluginFromValues)}
+                        >
+                            <Field
+                                name={`${name}.config.request_headers`}
+                                type="text"
+                                edit={edit}
+                                value={() => getValues(['config', 'request_headers'])(plugin)}
+                                placeholder={SETUP.placeholders.cors.request_headers}
+                                options={optionsTransformer(allValues('request_headers'))}
+                                component={TagSelect}
+                            />
+                        </CopyToClipboard>
                         <Hint>Value(s) for the Access-Control-Allow-Headers header.</Hint>
                     </Row>
                     <Row col>
